Moving cbor to encoding
