+++ rust/Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "tw_any_coin",
     "tw_bitcoin",
-    "tw_cbor",
     "tw_coin_entry",
     "tw_coin_registry",
     "tw_encoding",
