--- rust/tw_cbor/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "tw_cbor"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-arbitrary = { version = "1", optional = true, features = ["derive"] }
-ciborium = "0.2.1"
-serde = { version = "1.0.136", features = ["derive"] }
-
-[dev-dependencies]
-serde_bytes = "0.11.2"
