--- rust/tw_cbor/fuzz/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "tw_cbor-fuzz"
-version = "0.0.0"
-publish = false
-edition = "2021"
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-libfuzzer-sys = { version = "0.4", features = ["arbitrary-derive"] }
-serde = { version = "1.0.136", features = ["derive"] }
-serde_bytes = "0.11.2"
-
-[dependencies.tw_cbor]
-path = ".."
-features = ["arbitrary"]
-
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
-[profile.release]
-debug = 1
-
-[[bin]]
-name = "tw_cbor_serialize_deserialize"
-path = "fuzz_targets/tw_cbor_serialize_deserialize.rs"
-test = false
-doc = false
