+++ rust/tw_encoding/fuzz/Cargo.toml
@@ -9,6 +9,8 @@ cargo-fuzz = true
 
 [dependencies]
 libfuzzer-sys = { version = "0.4", features = ["arbitrary-derive"] }
+serde = { version = "1.0.136", features = ["derive"] }
+serde_bytes = "0.11.12"
 
 [dependencies.tw_encoding]
 path = ".."
@@ -30,5 +32,11 @@ doc = false
 [[bin]]
 name = "base_decode"
 path = "fuzz_targets/base_decode.rs"
+test = false
+doc = false
+
+[[bin]]
+name = "cbor_encode_decode"
+path = "fuzz_targets/cbor_encode_decode.rs"
 test = false
 doc = false
