+++ rust/tw_encoding/fuzz/fuzz_targets/cbor_encode_decode.rs
@@ -4,7 +4,7 @@ use libfuzzer_sys::{arbitrary, fuzz_target};
 
 use serde::{Deserialize, Serialize};
 
-use tw_cbor::{deserialize, serialize};
+use tw_encoding::cbor::{decode, encode};
 
 #[derive(arbitrary::Arbitrary, Serialize, Deserialize, PartialEq, Debug)]
 struct User {
@@ -17,7 +17,7 @@ struct User {
 }
 
 fuzz_target!(|se_user: User| {
-    let serialized = serialize(&se_user).unwrap();
-    let de_user = deserialize::<User>(&serialized).unwrap();
+    let serialized = encode(&se_user).unwrap();
+    let de_user = decode::<User>(&serialized).unwrap();
     assert_eq!(se_user, de_user);
 });
