+++ rust/tw_encoding/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod base32;
 pub mod base58;
 pub mod base64;
+pub mod cbor;
 pub mod ffi;
 pub mod hex;
